+++ LivingWater/src/main/java/com/livingwater/services/impl/DispenserServiceImpl.java
@@ -4,6 +4,7 @@ import com.livingwater.dao.CustomerDao;
 import com.livingwater.dao.DispenserDao;
 import com.livingwater.entities.Customer;
 import com.livingwater.entities.Dispenser;
+import com.livingwater.entities.User;
 import com.livingwater.services.DispenserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -29,28 +30,39 @@ public class DispenserServiceImpl implements DispenserService {
     private CustomerDao customerDao;
 
     public ModelAndView addDispenser(HttpServletRequest request, HttpServletResponse response) {
-        ModelAndView view = new ModelAndView("customer-dispensers");
+        ModelAndView view;
 
-        String dispenser_id = request.getParameter("dispenser_id");
-        int customer_id = Integer.parseInt(request.getParameter("customer_id").trim());
+        User user1 = (User) request.getSession().getAttribute("session_login_user");
 
-        Customer customer = customerDao.getCustomer(customer_id);
+        if (user1 == null) {
 
-        Dispenser dispenser = new Dispenser();
+            view = new ModelAndView("login");
+        } else {
 
-        dispenser.setSerialNumber(dispenser_id);
-        dispenser.setCustomer(customer);
+            view = new ModelAndView("customer-dispensers");
 
-        dispenserDao.create(dispenser);
+            String dispenser_id = request.getParameter("dispenser_id");
+            int customer_id = Integer.parseInt(request.getParameter("customer_id").trim());
 
-        System.out.println("---------------CREATE DISPENSER");
+            Customer customer = customerDao.getCustomer(customer_id);
 
-        List<Dispenser> dispenserList = getAllDispenserFromCustomer(customer_id);
+            Dispenser dispenser = new Dispenser();
 
-        if (dispenserList != null) {
-            view.addObject("dispenserList", dispenserList);
+            dispenser.setSerialNumber(dispenser_id);
+            dispenser.setCustomer(customer);
+
+            dispenserDao.create(dispenser);
+
+            System.out.println("---------------CREATE DISPENSER");
+
+            List<Dispenser> dispenserList = getAllDispenserFromCustomer(customer_id);
+
+            if (dispenserList != null) {
+                view.addObject("dispenserList", dispenserList);
+            }
+            view.addObject("customer", customer);
         }
-        view.addObject("customer", customer);
+
         return view;
     }
 
@@ -61,19 +73,30 @@ public class DispenserServiceImpl implements DispenserService {
     }
 
     public ModelAndView deleteDispenser(Integer id, HttpServletRequest request, HttpServletResponse response) {
-        ModelAndView view = new ModelAndView("customer-dispensers");
+        ModelAndView view;
 
-        Integer customer_id = Integer.parseInt(String.valueOf(request.getSession().getAttribute("session_customer_id")));
+        User user1 = (User) request.getSession().getAttribute("session_login_user");
 
-        Dispenser dispenser = new Dispenser();
-        dispenser.setSerialNumber(String.valueOf(id));
-        dispenserDao.delete(dispenser);
+        if (user1 == null) {
 
-        List<Dispenser> dispenserList = getAllDispenserFromCustomer(customer_id);
+            view = new ModelAndView("login");
+        } else {
 
-        if (dispenserList != null) {
-            view.addObject("dispenserList", dispenserList);
+            view = new ModelAndView("customer-dispensers");
+
+            Integer customer_id = Integer.parseInt(String.valueOf(request.getSession().getAttribute("session_customer_id")));
+
+            Dispenser dispenser = new Dispenser();
+            dispenser.setSerialNumber(String.valueOf(id));
+            dispenserDao.delete(dispenser);
+
+            List<Dispenser> dispenserList = getAllDispenserFromCustomer(customer_id);
+
+            if (dispenserList != null) {
+                view.addObject("dispenserList", dispenserList);
+            }
         }
+
         return view;
     }
 }
